+++ ultracache/__init__.py
@@ -1,9 +1,1 @@
-from django.conf import settings
 
-
-SETTINGS = {}
-try:
-    SETTINGS.update(settings.ULTRACACHE)
-except AttributeError:
-    # No overrides
-    pass