+++ ultracache/tests/viewsets.py
@@ -1,3 +1,5 @@
+import django
+
 from rest_framework import viewsets, serializers
 
 from ultracache.tests.models import DummyModel
@@ -7,6 +9,8 @@ class DummySerializer(serializers.ModelSerializer):
 
     class Meta:
         model = DummyModel
+        if not django.get_version().startswith("1.6"):
+            fields = "__all__"
 
 
 class DummyViewSet(viewsets.ModelViewSet):