+++ ultracache/decorators.py
@@ -20,10 +20,10 @@ def cached_get(timeout, *params):
             # The type of the request gets muddled when using a function based
             # decorator. We must use a function based decorator so it can be
             # used in urls.py.
-            request = getattr(view_or_request, 'request', view_or_request)
+            request = getattr(view_or_request, "request", view_or_request)
 
             # If request not GET or HEAD never cache
-            if request.method.lower() not in ('get', 'head'):
+            if request.method.lower() not in ("get", "head"):
                 return view_func(view_or_request, *args, **kwargs)
 
             # If request contains messages never cache
@@ -47,14 +47,14 @@ def cached_get(timeout, *params):
             ))):
                 li.append(request.get_full_path())
 
-            if 'django.contrib.sites' in settings.INSTALLED_APPS:
+            if "django.contrib.sites" in settings.INSTALLED_APPS:
                 li.append(settings.SITE_ID)
 
             # Pre-sort kwargs
             keys = kwargs.keys()
             keys.sort()
             for key in keys:
-                li.append('%s,%s' % (key, kwargs[key]))
+                li.append("%s,%s" % (key, kwargs[key]))
 
             # Extend cache key with custom variables
             for param in params:
@@ -62,27 +62,27 @@ def cached_get(timeout, *params):
                     param = str(param)
                 li.append(eval(param))
 
-            hashed = md5.new(':'.join([str(l) for l in li])).hexdigest()
-            cache_key = 'ucache-get-%s' % hashed
+            hashed = md5.new(":".join([str(l) for l in li])).hexdigest()
+            cache_key = "ucache-get-%s" % hashed
             cached = cache.get(cache_key, None)
             if cached is None:
                 # The get view as outermost caller may bluntly set _ultracache
                 request._ultracache = []
                 response = view_func(view_or_request, *args, **kwargs)
-                content = getattr(response, 'rendered_content', None) \
-                    or getattr(response, 'content', None)
+                content = getattr(response, "rendered_content", None) \
+                    or getattr(response, "content", None)
                 if content is not None:
-                    headers = getattr(response, '_headers', {})
+                    headers = getattr(response, "_headers", {})
                     cache.set(
                         cache_key,
-                        {'content': content, 'headers': headers},
+                        {"content": content, "headers": headers},
                         timeout
                     )
                     cache_meta(request, cache_key)
             else:
-                response = HttpResponse(cached['content'])
+                response = HttpResponse(cached["content"])
                 # Headers has a non-obvious format
-                for k, v in cached['headers'].items():
+                for k, v in cached["headers"].items():
                     response[v[0]] = v[1]
 
             return response