+++ ultracache/templatetags/ultracache_tags.py
@@ -2,7 +2,6 @@ from django import template
 from django.utils.translation import ugettext as _
 from django.utils.functional import Promise
 from django.templatetags.cache import CacheNode
-from django.template import resolve_variable
 from django.template.base import VariableDoesNotExist
 from django.core.cache import cache
 from django.core.cache.utils import make_template_fragment_key
@@ -32,32 +31,32 @@ class UltraCacheNode(CacheNode):
             expire_time = self.expire_time_var.resolve(context)
         except VariableDoesNotExist:
             raise TemplateSyntaxError(
-                '"cache" tag got an unknown variable: %r' % self.expire_time_var.var
+                "ultracache tag got an unknown variable: %r" % self.expire_time_var.var
             )
         try:
             expire_time = int(expire_time)
         except (ValueError, TypeError):
             raise TemplateSyntaxError(
-                '"cache" tag got a non-integer timeout value: %r' % expire_time
+                "ultracache tag got a non-integer timeout value: %r" % expire_time
             )
 
-        request = context['request']
+        request = context["request"]
 
         # If request not GET or HEAD never cache
-        if request.method.lower() not in ('get', 'head'):
+        if request.method.lower() not in ("get", "head"):
             return self.nodelist.render(context)
 
         # Set a list on the request. Django's template rendering is recursive
         # and single threaded so we can use a list to keep track of contained
         # objects.
-        if not hasattr(request, '_ultracache'):
-            setattr(request, '_ultracache', [])
+        if not hasattr(request, "_ultracache"):
+            setattr(request, "_ultracache", [])
             start_index = 0
         else:
             start_index = len(request._ultracache)
 
         vary_on = []
-        if 'django.contrib.sites' in settings.INSTALLED_APPS:
+        if "django.contrib.sites" in settings.INSTALLED_APPS:
             vary_on.append(str(settings.SITE_ID))
 
         for var in self.vary_on:
@@ -78,21 +77,21 @@ class UltraCacheNode(CacheNode):
         else:
             # A cached result was found. Set tuples in _ultracache manually so
             # outer template tags are aware of contained objects.
-            for tu in cache.get(cache_key + '-objs', []):
+            for tu in cache.get(cache_key + "-objs", []):
                 request._ultracache.append(tu)
 
         return value
 
 
-@register.tag('ultracache')
+@register.tag("ultracache")
 def do_ultracache(parser, token):
     """Based on Django's default cache template tag"""
-    nodelist = parser.parse(('endultracache',))
+    nodelist = parser.parse(("endultracache",))
     parser.delete_first_token()
     tokens = token.split_contents()
     if len(tokens) < 3:
-        raise TemplateSyntaxError("'%r' tag requires at least 2 arguments." % tokens[0])
+        raise TemplateSyntaxError(""%r" tag requires at least 2 arguments." % tokens[0])
     return UltraCacheNode(nodelist,
         parser.compile_filter(tokens[1]),
-        tokens[2], # fragment_name can't be a variable.
+        tokens[2], # fragment_name can"t be a variable.
         [parser.compile_filter(token) for token in tokens[3:]])