+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 next
 ----
 #. Add caching for Django Rest Framework viewsets.
+#. Django 1.10 compatibility.
 
 1.9.1
 -----